+++ common/src/main/java/com/chefmooon/frightsdelight/common/advancement/CandyBasketShuffleTrigger.java
@@ -1,0 +1,42 @@
+package com.chefmooon.frightsdelight.common.advancement;
+
+import com.chefmooon.frightsdelight.FrightsDelight;
+import com.google.gson.JsonObject;
+import net.minecraft.advancements.critereon.AbstractCriterionTriggerInstance;
+import net.minecraft.advancements.critereon.ContextAwarePredicate;
+import net.minecraft.advancements.critereon.DeserializationContext;
+import net.minecraft.advancements.critereon.SimpleCriterionTrigger;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerPlayer;
+
+public class CandyBasketShuffleTrigger extends SimpleCriterionTrigger<CandyBasketShuffleTrigger.TriggerInstance> {
+    private static final ResourceLocation ID = new ResourceLocation(FrightsDelight.MOD_ID, "use_baking_mat");
+    @Override
+    protected CandyBasketShuffleTrigger.TriggerInstance createInstance(JsonObject json, ContextAwarePredicate predicate, DeserializationContext deserializationContext) {
+        return new TriggerInstance(predicate);
+    }
+
+    @Override
+    public ResourceLocation getId() {
+        return ID;
+    }
+
+    public void trigger(ServerPlayer player) {
+        this.trigger(player, TriggerInstance::test);
+    }
+
+    public static class TriggerInstance extends AbstractCriterionTriggerInstance
+    {
+        public TriggerInstance(ContextAwarePredicate player) {
+            super(CandyBasketShuffleTrigger.ID, player);
+        }
+
+        public static TriggerInstance simple() {
+            return new TriggerInstance(ContextAwarePredicate.ANY);
+        }
+
+        public boolean test() {
+            return true;
+        }
+    }
+}