+++ fabric/src/main/java/com/chefmooon/frightsdelight/common/block/AbstractCandyBasketBlockImpl.java
@@ -1,10 +1,13 @@
 package com.chefmooon.frightsdelight.common.block;
 
+import com.chefmooon.frightsdelight.common.advancement.CandyBasketShuffleTrigger;
 import com.chefmooon.frightsdelight.common.block.entity.CandyBasketBlockEntity;
 import com.chefmooon.frightsdelight.common.data.types.CandyBasketType;
+import com.chefmooon.frightsdelight.common.registry.FrightsDelightAdvancements;
 import com.chefmooon.frightsdelight.common.registry.FrightsDelightSounds;
 import net.fabricmc.fabric.api.event.player.UseBlockCallback;
 import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
@@ -37,6 +40,7 @@ public class AbstractCandyBasketBlockImpl extends AbstractCandyBasketBlock {
                         Supplier<SoundEvent> shuffleSound = level.getBlockState(pos).getBlock() instanceof AbstractCandyBasketBlockImpl candyBasketBlock ? candyBasketBlock.getCandyBasketType().getShuffleSound() : null;
                         if (shuffleSound != null)
                             level.playLocalSound(pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, shuffleSound.get(), SoundSource.PLAYERS, Math.min(0.3f + 0.04f * candyBasketBlockEntity.getFilledSlotsCount(), 1.0f), 1.0f, false);
+                        if (player instanceof ServerPlayer serverPlayer) FrightsDelightAdvancements.CANDY_BASKET_SHUFFLE_TRIGGER.trigger(serverPlayer);
                         return InteractionResult.SUCCESS;
                     }
                 } else {